projetos/creditcard_fraud_detection/creditcard/src/requirements.txt
# Python 3.8
pandas>=1.3.3,<1.6.0
scikit-learn>=1.0,<1.3.0
matplotlib>=3.4.3,<4.0
seaborn>=0.11.2,<0.12
pyYAML==5.4.1
imbalanced-learn>=0.8.0,<0.10
xgboost>=1.5.2,<2.0
lightgbm>=3.3.1,<4.0
catboost>=1.2.7,<2.0
flask>=2.0,<3.0
pyspark==3.5.2
kafka-python==2.0.2
# boto3==1.18.0
# streamlit
# numpy>=1.22,<1.24
# dask[dataframe]>=2021.9.1
# pyarrow==4.0.1
# nltk>=3.6.3,<4.0
# gensim>=4.0.1,<5.0
# joblib>=1.0.1,<2.0
# pytest>=6.2.4,<7.0
# cython>=0.29.21,<0.30
# pandas-profiling>=3.1.0,<4.0
# setuptools>=49.6.0
# six>=1.15.0
# smart-open>=6.0.0
# click>=8.0.1,<9.0
# wordcloud>=1.8.1,<2.0
# Pillow>=9.3.0,<10.0
# textblob>=0.15.3,<0.16
# emoji>=2.6.0
# symspellpy>=6.7.8
# spacy>=3.7.6,<4.0
# vaderSentiment>=3.0.0
# torch>=2.0.0,<2.1.0
# transformers>=4.46.3,<5.0
# tensorflow>=2.12.0,<2.13.0
# tensorflow-estimator>=2.12.0,<2.13.0
# tensorflow-intel>=2.12.0,<2.13.0
# typing-extensions!=4.7.0,>=4.6.0
# accelerate>=0.26.0,<0.27.0
# langdetect
# keras>=2.12.0,<2.13.0
# fasttext>=0.9.2,<1.0
# scikit-optimize>=0.10.2,<0.11.0
# wrapt>=1.11.0,<1.15
# markupsafe>=2.1.1,<3.0
# dask-ml>=2022.1.1,<2023.0
